main.py
from agent.prompts import *
from agent.states import Plan, TaskPlan, CoderState
from dotenv import load_dotenv
from langchain_core.globals import set_debug, set_verbose
_ = load_dotenv()

from langgraph.constants import END
from langgraph.graph import StateGraph
from langgraph.prebuilt import create_react_agent
from agent.tools import write_file, read_file, get_current_directory, list_files, init_project_root

from langchain_groq import ChatGroq
llm = ChatGroq(model="openai/gpt-oss-120b")

set_debug(True)
set_verbose(True)




def planner_agent(state: dict) -> dict:
    user_prompt = state["user_prompt"]
    resp = llm.with_structured_output(Plan).invoke(planner_prompt(user_prompt))
    if resp is None:
        raise ValueError("Planner didn't returned a valid response.")

    return { "plan" : resp }

def architect_agent(state: dict) -> dict:
    plan : Plan = state["plan"]
    resp = llm.with_structured_output(TaskPlan).invoke(architect_prompt(plan))
    if resp is None:
        raise ValueError("Architect didn't returned a valid response.")
    
    resp.plan = plan
    
    return {"task_plan" : resp}

def coder_agent(state: dict) -> dict:
    """LangGraph tool-using coder agent."""
    coder_state: CoderState = state.get("coder_state")
    if coder_state is None:
        coder_state = CoderState(task_plan=state["task_plan"], current_step_idx=0)

    steps = coder_state.task_plan.implementation_steps
    if coder_state.current_step_idx >= len(steps):
        return {"coder_state": coder_state, "status": "DONE"}

    current_task = steps[coder_state.current_step_idx]
    existing_content = read_file.run(current_task.filepath)

    system_prompt = coder_system_prompt()
    user_prompt = (
    	f"Task: {current_task.task_description}\n"
    	f"File: {current_task.filepath}\n"
    	f"Existing content: \n{existing_content}\n"
    	"Use write_file(path, content) to save your changes."
    )
    coder_tools = [read_file, write_file, list_files, get_current_directory]
    react_agent = create_react_agent(llm, coder_tools)
    react_agent.invoke({"messages": [{"role": "system", "content": system_prompt},
								    {"role": "user", "content": user_prompt}]})

    coder_state.current_step_idx += 1
    return {"coder_state": coder_state}

graph = StateGraph(dict)

graph.add_node("planner", planner_agent)
graph.add_node("architect", architect_agent)
graph.add_node("coder", coder_agent)

graph.add_edge( "planner", "architect" )
graph.add_edge("architect", "coder")
graph.add_conditional_edges(
	"coder",
	lambda s: "END" if s.get("status") == "DONE" else "coder",
	{"END": END, "coder": "coder"}
	)
graph.set_entry_point("planner")
agent = graph.compile()

if __name__ == "__main__":
	init_project_root()
	result = agent.invoke({"user_prompt" : "Create a simple to do list web application"},
	{"recursion_limit": 100})
	
	print("Final State:", result)